[eclipse/xtext-eclipse#463] Editor stub is now only generated for Xbase languages.
- removed EditorFragment2, put its functionality into
XbaseGeneratorFragment2
- moved the binding of the editor from ImplicitFragment to
XbaseGeneratorFragment2
