+++ org.eclipse.xtext.xtext.generator/src/org/eclipse/xtext/xtext/generator/StandardLanguage.xtend
@@ -38,7 +38,6 @@ import org.eclipse.xtext.xtext.generator.validation.ValidatorFragment2
 import org.eclipse.xtext.xtext.generator.web.WebIntegrationFragment
 import org.eclipse.xtext.xtext.generator.xbase.XbaseGeneratorFragment2
 import org.eclipse.xtext.xtext.generator.xbase.XtypeGeneratorFragment2
-import org.eclipse.xtext.xtext.generator.ui.editor.EditorFragment2
 
 /**
  * This specialization of the {@link XtextGeneratorLanguage} adds all the standard generator fragments
@@ -117,8 +116,6 @@ import org.eclipse.xtext.xtext.generator.ui.editor.EditorFragment2
 	
 	SimpleProjectWizardFragment2 newProjectWizardForEclipse = new SimpleProjectWizardFragment2
 	
-	EditorFragment2 editor = new EditorFragment2
-	
 	new() {
 		try {
 			class.classLoader.loadClass("org.eclipse.xtext.xbase.XbaseRuntimeModule")
@@ -174,7 +171,6 @@ import org.eclipse.xtext.xtext.generator.ui.editor.EditorFragment2
 		fragments += ideaPlugin
 		fragments += webSupport
 		fragments += newProjectWizardForEclipse
-		fragments += editor
 		fragments
 	}
